Moved Rule IDs to prepared query values array
